Converts a set of (x,y) coords into a binary image
This program will take the (x,y) coords of a room generated by the AFX-10 and convert them into a binary image that can be used for computer vision purposes on the robot. With this, it will be possible to detect where a wall is in the image based on if a pixel is filled or not. The resolution of the generated image can be adjusted.