Basic robot path planning algorithm
This code is used to navigate through the grid space that was created with Coords2Image. The goal is to start at the top left corner, and end at the bottom right corner. This code will make decisions on where to travel next based on the locations of obstacles around the robot. The code will plot the path the robot takes step by step as it makes its way towards the goal.